perf(shell): finish narrowing instance/clock subscriptions + SW precache and react-scan fixes
- IpodWidget: playback clock (~20Hz) moved into leaf progress-fill and
  time-label components; the widget body no longer re-renders per tick
- Windows taskbar (MenuBar): reuse MacDock's signature-keyed instances
  snapshot so geometry/focus writes don't re-render the taskbar
- AppMenu: boolean hasMinimizedInstances selector; bulk hide/show handlers
  read instances via getState() at click time
- ExposeView: instances subscription gated on overlay visibility
- Extract shared dockInstancesSnapshot module (was private to MacDock)
- Workbox: exclude pusher-js and react-player chunks from SW precache
  (network-dependent features, useless offline)
- ReactScanDebug: only download the react-scan chunk when debug mode is on
