+++ src/components/ReactScanDebug.tsx
@@ -10,13 +10,23 @@ export function ReactScanDebug() {
   const debugMode = useDisplaySettingsStore((s) => s.debugMode);
 
   useEffect(() => {
+    // Only download the react-scan chunk when debug mode is actually on —
+    // an unconditional import() fetched it on every session even though
+    // scanning was disabled.
+    if (!debugMode) return;
     import("react-scan").then(({ scan }) => {
       scan({
-        enabled: debugMode,
-        showToolbar: debugMode,
-        dangerouslyForceRunInProduction: debugMode,
+        enabled: true,
+        showToolbar: true,
+        dangerouslyForceRunInProduction: true,
       });
     });
+    return () => {
+      // Turn scanning off if debug mode is disabled after the module loaded.
+      import("react-scan").then(({ scan }) => {
+        scan({ enabled: false, showToolbar: false });
+      });
+    };
   }, [debugMode]);
 
   return null;
